add logistic error. Find error in least squares
OLS: The least squares method seems to have a problem with singular matrices. Refer to the code for more detail.

SGD: When the number of considered datapoints increases to a reasonable size, the SGD performs terribly. I guess this is expected because we have to implement it with batch-size 1, so we'll leave the implementation as-is. The implementation still needs to be tested just in case. If we ever want to actually use SGD, we'll have to increase the batch-size for sure.